requirements.txt
strands-agents
strands-agents-tools
strands_agents_builder
beautifulsoup4
colorama